requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.1
optuna==3.2.0
fastapi
uvicorn
pillow
python-multipart